Feature swerve drive robot

+++ src/control/robot_base/src/speed_actuator_group.cpp
@@ -7,20 +7,20 @@
  * Copyright (c) 2024 Ruixiang Du (rdu)
  */
 
-#include "robot_base/speed_actuator_group.hpp"
+#include "robot_base/actuator/speed_actuator_group.hpp"
 
 namespace xmotion {
-void SpeedActuatorGroup::SetSpeed(int32_t rpm) {
+void SpeedActuatorGroup::SetSpeed(float rpm) {
   for (auto &actuator : actuators_) actuator->SetSpeed(rpm);
 }
 
-int32_t SpeedActuatorGroup::GetSpeed() {
+float SpeedActuatorGroup::GetSpeed() {
   int64_t average_speed = 0;
   for (auto &actuator : actuators_) average_speed += actuator->GetSpeed();
   return average_speed / actuators_.size();
 }
 
-void SpeedActuatorGroup::ApplyBrake(double brake) {
+void SpeedActuatorGroup::ApplyBrake(float brake) {
   for (auto &actuator : actuators_) actuator->ApplyBrake(brake);
 }
